separate test_details out of component_report
